fix(check): Avoid instantiating higher ranked types in function arguments
